requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
boltons==21.0.0
bracex==2.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-option-group==0.5.6
colorama==0.4.6
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.0
face==22.0.0
frozenlist==1.4.1
glom==22.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
mysql-connector-python==8.3.0
nltk==3.8.1
numpy==1.26.4
openai==1.21.2
packaging==23.2
peewee==3.17.1
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
python-dotenv==1.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.1.post1
scipy==1.13.0
semgrep==1.63.0
sniffio==1.3.1
soupsieve==2.5
threadpoolctl==3.4.0
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
wcmatch==8.5.1
yarl==1.9.4
